+++ src/main/java/assemble/api/club/domain/Club.java
@@ -4,11 +4,11 @@ import assemble.api.board.domain.Board;
 import assemble.api.chat.domain.Chat;
 import assemble.api.club.domain.enums.ClubStatus;
 import assemble.api.club.domain.enums.DifficultyLevel;
+import assemble.api.club.domain.mapping.ClubJoinRequest;
 import assemble.api.club.domain.mapping.MemberClub;
 import assemble.api.club.dto.ClubRequestDTO;
 import assemble.api.global.base.BaseEntity;
 import assemble.api.member.domain.enums.InterestCategory;
-import assemble.api.member.domain.mapping.ClubJoinRequest;
 import assemble.api.member.domain.mapping.MemberLikesClub;
 import assemble.api.notice.domain.Notice;
 import assemble.api.schedule.domain.Schedule;
