+++ src/main/java/assemble/api/member/domain/Member.java
@@ -7,7 +7,7 @@ import assemble.api.comment.domain.Comment;
 import assemble.api.global.base.BaseEntity;
 import assemble.api.likes.domain.Likes;
 import assemble.api.member.domain.enums.InterestCategory;
-import assemble.api.member.domain.mapping.ClubJoinRequest;
+import assemble.api.club.domain.mapping.ClubJoinRequest;
 import assemble.api.club.domain.mapping.MemberClub;
 import assemble.api.member.domain.mapping.MemberLikesClub;
 import assemble.api.member.domain.mapping.MemberSchedule;
